requirements.txt
# Core ML Libraries
torch==2.5.0
torchvision
torchaudio
transformers==4.44.2
accelerate==0.34.2
bitsandbytes==0.43.3
peft
trl<0.9.0
xformers
unsloth[colab-new]

# Text Processing
nltk
textstat
python-docx
PyPDF2
tokenizers==0.19.1
tiktoken==0.4.0
regex==2024.7.24

# Data Science & Utils
numpy==1.26.3
pandas
scikit-learn
datasets
wandb
tqdm==4.66.5

# File Handling & System
fsspec==2024.2.0
filelock==3.13.1
requests==2.32.3
urllib3==2.2.2
psutil==6.0.0

# Development & Build Tools
cmake==3.30.2
ninja==1.11.1.1
packaging==24.1

# Other Dependencies
safetensors==0.4.5
huggingface-hub==0.24.6
pillow==10.2.0
sympy==1.13.1
networkx==3.2.1
jinja2==3.1.3
pyyaml==6.0.2 